requirements.txt
PySide6==6.8.1.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
black==24.10.0
python-tcxparser==2.3.0